% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCellData.R
\name{xCellData}
\alias{xCellData}
\title{XCellData: 489 cell type gene signatures}
\usage{
xCellData()
}
\value{
The gene sets as a \code{\link{unisets-package}} \code{\link{BaseSets}}.
}
\description{
XCellData: 489 cell type gene signatures
}
\examples{
xCellData()
}
